pyproject.toml
[project]
name = "kustomanager"
description = "Project automation for Kustomization-based GitOps repositories"
authors = [{ name = "Brian Lu", email = "me@greencappuccino.net" }]
dependencies = [
    "click>=8.1.7",
    "ruamel-yaml>=0.18.5",
    "gitignore-parser>=0.1.9",
    "Jinja2>=3.1.3",
]
requires-python = "==3.12.*"
readme = "README.md"
license = { text = "MIT" }
dynamic = ["version"]

[project.scripts]
kustomanager = "kustomanager.__main__:main"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
package-type = "library"

[tool.pdm.version]
source = "file"
path = "kustomanager/meta.py"

[tool.pdm.build]
includes = ["kustomanager/"]

[tool.pdm.dev-dependencies]
dev = ["pre-commit>=3.6.0", "yamllint>=1.33.0", "mypy>=1.8.0", "ruff>=0.1.11"]

[tool.pdm.scripts]
post_install = { cmd = ["pre-commit", "install"] }

[tool.ruff]
select = ["E4", "E7", "E9", "F", "B", "I", "Q"]
